+++ NEWS.md
@@ -21,6 +21,7 @@
 ## New features
 
 - Introduce a new experimental API for generating large plans (#674). It is inspired by the proposed DSL (#233) and the `target()` function, two of @krlmlr's ideas. Uses custom fields `transform` and `group` in `target()` within `drake_plan()`. New arguments `transform` and `trace` in `drake_plan()` provide some customization.
+- Add a new function `transform_plan()` to do the transformations in the above API with an existing plan. Useful for combining multiple plans that were generated with `drake_plan(transform = FALSE)` so you can generate multiple un-transformed plans, combine them, and then transform the aggregate plan.
 - Implement a `lock_envir` argument to safeguard reproducibility. See [this thread](https://github.com/ropensci/drake/issues/615#issuecomment-447585359) for a demonstration of the problem solved by `make(lock_envir = TRUE)`. More discussion: #619, #620.
 - The new `from_plan()` function allows the users to get custom columns of the plan. Changes to these custom columns do not invalidate targets, however, so be careful.